Found the error.
Problem was when we try to skip yolo model checkings by using a skipping_frame variable for unknown reason deepsort unable to confirm any of the tracks it recieved. Problem goes away when we removed the yolo limiting.

Cross Camera system is also destroyed due to trying for debug the problem. Need to implement that again